requirements.txt
flask==2.3.3
flask-session==0.5.0
requests==2.31.0
python-telegram-bot==20.7
instaloader==4.10.3
schedule==1.2.0
python-dotenv==1.0.0
werkzeug==2.3.7
